requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
eventlet==0.36.1
traci==1.21.0
sumolib==1.21.0
lightgbm==4.4.0
scikit-learn==1.5.1
networkx==3.3
osmnx==1.9.4
pandas==2.2.2
numpy==1.26.4
shapely==2.0.5
matplotlib==3.9.1
joblib==1.4.2
python-dotenv==1.0.1
requests==2.32.3
lxml==5.3.0
# Ensemble additions
catboost>=1.2
torch>=2.2.0
torch_geometric>=2.5.0
